Delete Route Refactored

+++ node-rest-api/controllers/feed.js
@@ -136,6 +136,12 @@ exports.editPost = (req, res, next) => {
 				throw error;
 			}
 
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
+
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -170,11 +176,23 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
 		.then((result) => {
-			console.log(result);
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
+		})
+		.then((result) => {
 			res.status(200).json({ message: 'Post Deleted' });
 		})
 		.catch((err) => {
